+++ rapidfuzz/distance/Prefix.hpp
@@ -10,28 +10,28 @@
 namespace rapidfuzz {
 
 template <typename InputIt1, typename InputIt2>
-int64_t prefix_distance(InputIt1 first1, InputIt1 last1, InputIt2 first2, InputIt2 last2,
-                        int64_t score_cutoff = std::numeric_limits<int64_t>::max())
+size_t prefix_distance(InputIt1 first1, InputIt1 last1, InputIt2 first2, InputIt2 last2,
+                       size_t score_cutoff = std::numeric_limits<size_t>::max())
 {
     return detail::Prefix::distance(first1, last1, first2, last2, score_cutoff, score_cutoff);
 }
 
 template <typename Sentence1, typename Sentence2>
-int64_t prefix_distance(const Sentence1& s1, const Sentence2& s2,
-                        int64_t score_cutoff = std::numeric_limits<int64_t>::max())
+size_t prefix_distance(const Sentence1& s1, const Sentence2& s2,
+                       size_t score_cutoff = std::numeric_limits<size_t>::max())
 {
     return detail::Prefix::distance(s1, s2, score_cutoff, score_cutoff);
 }
 
 template <typename InputIt1, typename InputIt2>
-int64_t prefix_similarity(InputIt1 first1, InputIt1 last1, InputIt2 first2, InputIt2 last2,
-                          int64_t score_cutoff = 0)
+size_t prefix_similarity(InputIt1 first1, InputIt1 last1, InputIt2 first2, InputIt2 last2,
+                         size_t score_cutoff = 0)
 {
     return detail::Prefix::similarity(first1, last1, first2, last2, score_cutoff, score_cutoff);
 }
 
 template <typename Sentence1, typename Sentence2>
-int64_t prefix_similarity(const Sentence1& s1, const Sentence2& s2, int64_t score_cutoff = 0)
+size_t prefix_similarity(const Sentence1& s1, const Sentence2& s2, size_t score_cutoff = 0)
 {
     return detail::Prefix::similarity(s1, s2, score_cutoff, score_cutoff);
 }
@@ -63,7 +63,7 @@ double prefix_normalized_similarity(const Sentence1& s1, const Sentence2& s2, do
 }
 
 template <typename CharT1>
-struct CachedPrefix : public detail::CachedSimilarityBase<CachedPrefix<CharT1>, int64_t, 0,
+struct CachedPrefix : public detail::CachedSimilarityBase<CachedPrefix<CharT1>, size_t, 0,
                                                           std::numeric_limits<int64_t>::max()> {
     template <typename Sentence1>
     explicit CachedPrefix(const Sentence1& s1_) : CachedPrefix(detail::to_begin(s1_), detail::to_end(s1_))
@@ -74,19 +74,18 @@ struct CachedPrefix : public detail::CachedSimilarityBase<CachedPrefix<CharT1>, 
     {}
 
 private:
-    friend detail::CachedSimilarityBase<CachedPrefix<CharT1>, int64_t, 0,
-                                        std::numeric_limits<int64_t>::max()>;
+    friend detail::CachedSimilarityBase<CachedPrefix<CharT1>, size_t, 0, std::numeric_limits<int64_t>::max()>;
     friend detail::CachedNormalizedMetricBase<CachedPrefix<CharT1>>;
 
     template <typename InputIt2>
-    int64_t maximum(const detail::Range<InputIt2>& s2) const
+    size_t maximum(const detail::Range<InputIt2>& s2) const
     {
-        return static_cast<int64_t>(std::max(s1.size(), s2.size()));
+        return std::max(s1.size(), s2.size());
     }
 
     template <typename InputIt2>
-    int64_t _similarity(detail::Range<InputIt2> s2, int64_t score_cutoff,
-                        [[maybe_unused]] int64_t score_hint) const
+    size_t _similarity(detail::Range<InputIt2> s2, size_t score_cutoff,
+                       [[maybe_unused]] size_t score_hint) const
     {
         return detail::Prefix::similarity(s1, s2, score_cutoff, score_cutoff);
     }
